Committing first crack at Java version of the events test.
Currently all parameters are hard-coded in the file, but the first
three phases work. You will likely have to up the heap size
for the VM when you run this (try -Xm512m).

The monitor.sh script is called at the start and end of each phase.

Working on getting Phase 4 and 5 in now, and reading from a properties
file instead..