AG4 Java Tutorials update.
TutorialExamples.java

A small change to example20() on the way to testing
geo subtypes.  This is an intermediate stage, but it
runs cleanly without calling attention to the changes.
I will check in a finished version in a day or so. BDC

Release Notes: No
Test Suite run? No.
BDC