rfe8812: Jena GraphMaker and Model compliance tests
Added AGGraphMakerTest and AGModelTest to exercise 27 tests for
Jena GraphMaker and Model compliance, along with supporting code
changes that allow 26 of them to pass.  The remaining test will
be addressed in a future commit.

Tests added for:  AGGraphMakerTest and AGModelTest for rfe8812
make test-suite run? no
make prepush run? yes (in agraph-java-client)

<release-note>
rfe8812: Jena compliance improvements

Added compliance tests for Jena GraphMaker and Model interfaces and
corrected several non-compliances.

Properly handle the case of getting a namespace uri for a prefix that
does not exist, and add support for expanding a QName.

Add GraphMaker support for creating anonymous graphs, listing graphs,
checking existence of a graph, and returning existing graph objects
rather than creating new ones when possible.

Throw AlreadyExistsException when attempting to create an existing
graph in strict mode, and throw a DoesNotExist exception when
attempting to remove or open a non-existent graph in strict mode.
</release-note>

Signed-off-by: Ahmon Dancy <dancy@franz.com>