bug19063: NPE in encodeJSON
In repairing bug19054 there was a regression that resulted in a NPE
when encodingJSON for statements in the null context.  This is now
fixed.

Tests added for: JenaTutorialExamples#example2 exercises bug19063.
make test-suite run? no
make prepush run? yes
ant tutorial run? yes
ant jen-tutorial run? yes

releases notes: no, this is a correction to the repair of bug19054.