Restore previous createRepository behavior - bug19066
Previously Catalog.createRepository would create the repository if it did not
exist, or open it otherwise in the Sesame client. When the override parameter
was added to the request, the default behavior of put on a repository changed.
This change adds the override='false' to all put repository requests to restore
the previous behavior for the Java Sesame client APIs.

This allows the events test to run unmodified, so no release notes are needed
as the external API has not changed.

Also added a small change to build.xml to give the events test JVM more
space to run using jvmargs.

Tests added for:     none
make test-suite run? yes, and test-stress-events

Signed-off-by: Ahmon Dancy <dancy@franz.com>