rfe9392: use accept-encoding gzip to get back compressed results
Get and Post requests now use the Accept-Encoding: gzip header to
request that responses be compressed, to improve performance when
results are large.

Modified both tutorial examples files to call closeAll prior to
deleteRepository in example1, as not doing so can result in orphaned
sessions on the toClose list.

Added a source code attachment for the HTTPClient 3.1 jar, to aid
in development and debugging.

Tests added for: rfe9392
make test-suite run? no
make prepush run? yes
tutorials run? yes

<release-note>
rfe9392: use accept-encoding gzip to get back compressed results

Get and Post requests now use the Accept-Encoding: gzip header to
request that responses be compressed, to improve performance when
results are large.

Modified both tutorial examples files to call closeAll prior to
deleteRepository in exampele1, as not doing so can result in orphaned
sessions on the toClose list.
</release-note>