bug19280 - Java tutorial update to example12()
The free-text indexing example in the Java
tutorial (both html and code) needed to be
updated.  I have improved it, but some of the
free-text functionality seems missing. I have
upgraded the example (and text) so that it
isn't actually *wrong* anymore, but it will
need additional work in the future. I have
left the bug issue open.

No release notes. The java examples ran. The
html file looked ok in the browser.

Change-Id: I8ab74aa763376446ace78cfd5855233dd64b625e