rfe9744: upgrade to Jena 2.6.2
Previously the Java client was using Jena 2.6.0; this change upgrades the client to
the latest version of Jena.

Tests added for:  none.
make prepush run? yes.
ant jena-tutorial run? yes.

<release-note>
rfe9744: upgrade to Jena 2.6.2

Previously the Java client was using Jena 2.6.0; this change upgrades the client to
the latest version of Jena.

</release-note>

Change-Id: Id4d5e46bb40d15dbacc6b98edcb87753bbcd50d8