bug19445: remove log4j reference from classpath template
The classpath template in bug19440 was referencing a log4j jar that is
not in the distro.  Testing for bug19440 passed because the log4j jar
was present in the workspace as an untracked file, but the nightly
build failed from a clean checkout.  This change removes the spurious
reference.

Tests added: no.
git status is clean.
make prepush run? yes
distro's TutorialExamples run? yes.
distro's JenaTutorialExamples run? yes.

releases notes: non-user-visible bug fix prior to release.

Change-Id: I01d569de90255c8205cf98b63e401cb710be91c1