bug19569: java client now helps delete orphaned saved queries
Preparing a Sesame AGQuery currently has the side effect of saving the
prepared query on the server in memory for the duration of the session.
A client app that prepares many queries during a session thus uses up
server memory.  If an AGQuery instance goes out of scope, the saved
query is no longer needed on the server and is currently only deleted
when the session expires.  The client library now tries to help delete
orphaned saved queries sooner by tracking when AGQuery instances are
garbage collected on the client side and issuing delete requests for
orphaned saved queries in subsequent calls to prepare query.

Test: see AGPrepareQuery.java
make prepush run
ant tutorial run
ant jenatutorial run

<release-note>
bug19569: java client now helps delete no-longer-needed saved queries

Preparing a Sesame AGQuery currently has the side effect of saving the
prepared query on the server in memory for the life of the session.
A client app that prepares many queries during a session thus pressures
server memory.  If an AGQuery instance goes out of scope, the saved
query is no longer needed on the server and is currently only deleted
when the session expires.  The client library now tries to help delete
orphaned saved queries sooner by tracking when AGQuery instances are
garbage collected on the client side, and then issuing delete requests
for orphaned saved queries.
</release-note>

Change-Id: Id3ff3ab679561d35ff78717ca1001cc4cbe29f0a

more on prepare

Change-Id: I99514b79500cfa119a6df81688c3becaca110790