bug19280 Small changes to Java tutorial example12()
The context is the Java API to freetext indexing.
Pursuant to bug19280, rfe9319, spr37288, I have made
some changes to example12, both in code and in the
tutorial.  One line of the tutorial was out of sync
with the code, which confused a 4.0.5 user. Further changes
to the Java API, the example code, and the tutorial
are expected before 4.1.

No release notes.

Change-Id: Ia81e9a6351f8db2cc6719c552eaa33f81bee3667