rfe9831 - Java tutorial: Size matters.
Alterations to the the Java tutorial,
and TutorialExamples.java example10(), to work in
examples of conn.size() reporting on specific subgraphs.

<release-note>
rfe9831 - Java tutorial, illustrate use of conn.size() with contexts.

Example10() of the Java tutorial has been extended to
show examples of the conn.size() method returning the
number of triples found in specific subgraphs.
</release-note>

Change-Id: I3c0db14ba7a76af3c2dd4d93c3aa375d763e4436