Separate sync phases from phases 1-5. Adjust log message for easy parsing.
Events.java:

- Change Log Messages for phase begin/end to be "Phase N Begin:" and
"Phase N End:" to make regexp matching of lines in the log file
simpler.

- Also, the call to Monitor.Stop() (i.e. the sync phase) was being
called in a location where it's time was included in the time for
Phases 1-5.  We now track the timing of sync phases separately, so
this was moved to below the timing calculation for each completed
phase.

monitor.sh:

- Emit datetime stamps around the sync phase invocation. We use 'Phase
0' to denote each sync phase.

Tests added for: no
make test-suite run? no

release notes: non-user-visible changes

Change-Id: Ic36774b027e278d9d3efd9617ec5b1282db395ad