rfe10143: support bulkMode in connections
<release-note>
rfe10143: support bulkMode in connections

With this change, connections can now be put into bulkMode.
When in bulkMode, data can be added/loaded more quickly, but
there is no guarantee of durability in the event of a crash.
Added method AGRepositoryConnection#setBulkMode to set the
connection's bulkMode (defaults to false), and added method
AGRepositoryConnection#isBulkMode to check the mode.
<release-note>

Added BulkModeTests to prepush tests
make prepush passes
ant tutorial runs

Change-Id: I5b134555e727e56895d376dbb293ad176b9bd2e0
Reviewed-on: https://gerrit.franz.com:9080/948
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>