rfe9436: session lifetime management
<release-note>
rfe9436: session lifetime management

With this change the lifetime (maximum idle time) of a session can
be set using the AGRepositoryConnection#setSessionLifetime method.
The default session lifetime (in effect when not set with the method
above) can be set with the AGHttpRepoClient.setDefaultSessionLifetime
method.  When none of these are set, the default session lifetime is
1 hour.
</release-note>

Added SessionLifetimeTests to prepush tests
make prepush passes
ant tutorial passes

Change-Id: I757d4b874cb5d2e71e8c9c4466992e5da2228f25