WIP: rfe10189: Java Interface for Stored Procedures
WORK IN PROGRESS - DO NOT USE

Proposed java API for calling stored procs for v4.3.

The java code provided with rfe10189 is included, but
JFK is planning to rewrite this serialization code.

Using the curl command to upload the fasl, caused the
error below, so I copied the fasl manually to the server.

[2010-11-09T13:14:26.858 p18710 E backend] Error loading script
'/tmp/mike/v4.2/settings/scripts/example.fasl':\nAttempt to fast load
a non-fasl file: #P"/tmp/mike/v4.2/settings/scripts/example.fasl"

The included test does invoke the stored proc, but it fails with the
error below.  Run the java test with 'ant test-temp'.

[2010-11-09T13:17:14.579 p18710 E http] the value of excl::x is nil,
which is not of type integer.

Change-Id: I6b735d76342473c8dae8b49a43638a769a5a9664