rfe10197: support for encodable namespaces and generated URIs
<release-note>
rfe10197: support for encodable namespaces and generated URIs

With this change, a namespace (viewed as a set of URIs) can now be
registered as encodable.  URIs in the namespace have local parts that
follow a declared format, allowing for efficient encoding and in some
cases generation of unique URIs in the namespace.  Once registered,
a conformant URI in the namespace is auto-encoded when added to the
store, resulting in a smaller string table and faster query speed.
</release-note>

Added EncodableNamespaceTests to prepush tests
make prepush passes
ant tutorial runs

Change-Id: I5d070f347fd085cd247cb3d3e856ed83ec5311a7