rfe10189: Java Interface for Stored Procedures
Experimental java API for calling stored procs.
New classes and methods are marked deprecated in the javadocs
because they may change in a future release.

The java code provided with rfe10189 is included,
and JKF's encoding code.

<release-note>
rfe10189: Java Interface for Stored Procedures

The stored proc feature and API are experimental,
and subject to change in a future release.
For this reason, the new methods are marked as deprecated.

The AllegroGraph server defines a new API for defining Stored
Procedures and they are installed like Custom Services.

The primary API is:
AGRepositoryConnection.callStoredProc(functionName, moduleName, args)

A low-level API is also exposed:
AGHttpRepoClient.callStoredProcEncoded()
AGSerializer.serializeAndEncode()
AGDeserializer.decodeAndDeserialize()
</release-note>

tests added:       to test encoding and stored proc call
tests run:         prepush

Depends on commit for rfe10189 in agraph:
https://gerrit.franz.com:9080/982

Change-Id: I6b735d76342473c8dae8b49a43638a769a5a9664
Reviewed-on: https://gerrit.franz.com:9080/964
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>