rfe10117 - implement run-time query index analysis
Tests for query analysis cannot be in the java-prepush because
the default analysis technique is to use :execute and the stores in
the prepush are too small to have any on disk chunks.

The plan is to move something like these tests to the events test.

tests added    : no
tests run      : java-prepush
performance    : no change
release-note   : none, internal testing change

Change-Id: Icdf2f9e6cbbcb143b0a71ded8c2223137d22b8d9