rfe9758: automated tests for client distro's tar.gz
Test that eclipse will build/compile the distributed
project.  Eclipse is not easily scriptable from the
command line, so an eclipse workspace was manually
created and the minimal files included with this commit
(instructions to repeat in build.xml).

New target 'ant test-dist' uses that workspace to build the
dist project.  This target requires the eclipse-jdt package
is installed (and eclipse is on PATH).  Because of this
dependency, test-dist is run with test-bigger, not prepush.

Note: started using a target dir that can be used for
all build operations.

Incidentally, removed useless execute mode from some files.

tests-added:  test-dist to test-bigger
tests-run:    lisp-bigger-test
performance:  no impact, test only
release-note: no impact, test only

Change-Id: I3f95c4cf66fb9105b2edf721ae3f2c53c9cc1146