rfe10256 support loadInitFile switch in java
Adds method to make use of the param in the http protocol.

ServerCodeTests uploads an initFile to verify the param works
and fails if the param is not set.

tests-added:     yes
tests-run:       prepush
performance:     no impact

<release-note>
rfe10256 support loadInitFile switch in java

Added method to AGRepositoryConnection:
setSessionLoadInitFile(boolean loadInitFile)

When a dedicated session is created, the initFile on the
AllegroGraph Server will be loaded if this param is true.

Javadocs on AGRepositoryConnection reorganized to describe
sessions.
</release-note>

Change-Id: If2ea4999cae1dc05013963cc265c8e5534e209b7
Reviewed-on: https://gerrit.franz.com:9080/1191
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Reviewed-by: John O'Rourke <john.orourke@franz.com>
Tested-by: Kevin Layer <layer@franz.com>