rfe10257 support script loading in sessions
Adds method to make use of the param in the http protocol.

ServerCodeTests uploads a script to verify the param works
and fails if the param is not set.

tests-added:     yes
tests-run:       prepush
performance:     no impact

<release-note>
rfe10257 support script loading in sessions

Added method to AGRepositoryConnection:
addSessionLoadScript(String scriptName)

When a dedicated session is created, the named scripts
on the AllegroGraph Server will be loaded.
</release-note>

Change-Id: I5d56931db7e7345ba6e25bf6825884262315c645
Reviewed-on: https://gerrit.franz.com:9080/1192
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Reviewed-by: John O'Rourke <john.orourke@franz.com>
Tested-by: Kevin Layer <layer@franz.com>