rfe10411: Streaming large query results
Adds new API as an option to stream query results.

tests-added:   re-run prepush with streaming enabled
tests-run:     test-stress

<release-notes>
rfe10411: Streaming large query results

Query results in the sparql/xml format use the Sesame libary for
parsing, which uses a SAX XML parser, builds an ArrayList, then
iterates through results.  This will cause OutOfMemoryError for
large results.  This applies to queries in both the sparql and
prolog languages.

Alternatively, TupleQuery.evaluate(handler) can be used where the
handler gets callback functions as results are parsed.  This is
faster and uses less memory, can handle any length results, but
the API is less convenient.  A new example has been added to the
tutorial to demonstrate this: TutorialExamples.example3a()

A new class AGStreamTupleQuery can be used to wrap a TupleQuery
so the TupleQueryResult will be streamed back.  This provides
the familiar iteration API rather than a callback, but the http
response is only parsed as the iterator is advanced.  The usage
of this API is explained further in the javadocs:
AGTupleQuery query = conn.prepareTupleQuery(QueryLanguage.SPARQL,
"SELECT ...");
query = new AGStreamTupleQuery(query);
TupleQueryResult results = query.evaluate();

There is also a new method in AGRepositoryConnection that will
cause the connection to automatically wrap all tuple queries to
use streaming:
AGRepositoryConnection.setStreamResults(boolean streamResults)

Instead of SAX, the AGStreamTupleQuery uses an XMLStreamReader
for parsing.  Initial performance results suggest that in
addition to being resistant to OutOfMemoryError this uses less
overall memory than the SAX implementation for small and medium
result sets (10-500).

This approach of streaming results has not yet been applied to
getStatements().
</release-notes>

<performance>
The default method of queries and results is unchanged, this is
an addition to the API, so there is no impact unless users make
an explicit change.

Initial performance tests show improvement in both time and memory
with the XMLStreamReader implementation for the spectrum of result
set sizes, 10 to 1 million.

However, the Events.java test shows little overall improvement, so
more tuning may be possible to see a benefit.
</performance>

Change-Id: I8436b24e0e9a508c4f16682b123cf086eea62cc1