rfe10505: use session to get new blankNodeIds
In Events.java: Instead of using shared backends for queries, and
dedicated for add and delete phases, use dedicated for all.

tests-added:   none
tests-run:     prepush

<release-note>
rfe10505: use session to get new blankNodeIds

In AGValueFactory, use the AGRepositoryConnection to retrieve
new blankNodeIds, using the session if there is one active.
Using the session can reduce load on the primary server.
If the ValueFactory was created by Repository instead of
RepositoryConnection, the old behavior remains.
</release-note>

Change-Id: If39bdc52ed4804e1bfb7e772a14e74e0f3aeb08c