Add http api for sync's and merges. Updated events test to use them.
Add documented method optimizeIndices().
Add undocumented methods ensureDBIdle() and forceCheckpoint().

Fix bug where useDedicatedSession() was not setting
usingDedicatedSession flag. This was preventing the ping() command
from working.

Slightly adjusts the call sequence in the events test to close tasks
before moving on to the next phase. Also adds a Pinger class to call
ping() on the primary dedicated session. This is necessary to prevent
session timeouts in the case when a phase may take longer than the
session lifetime. A ping is sent every (sessionLifetime / 2) seconds.

make test-suite run? no

releases notes: user-visible API/code change(s)

<release-note>
optimizeIndices is now a method on AGRepositoryConnection instances.

This tells the server to try and optimize the indices for this
store. Takes a boolean wait argument, which indicates whether the HTTP
request should return right away or whether it should wait for the
operation to complete. Also takes an optional level argument (defaults
to 1), which determines how much work should be done. See the index
documentation
(http://www.franz.com/agraph/support/documentation/current/triple-index.html)
for the meaning of the different levels.
</release-note>

Change-Id: Iba89de6bf739e742cf23a75fce4bf7324f705117