bug20272: NPE when no response stream is available
<release-note>
bug20272: NPE when no response stream is available

Previously, the Java client's HTTP response handler could throw an
opaque NullPointerExeption if no response stream is available (this
might occur due to a server crash).  With this change, the response
handler checks for a null stream and throws an exception advising
the user that no response stream is available and to inspect the
server log file for possible explanations.
</release-note>

make prepush passes

Change-Id: Ie29cae3ad0b0069ec55d53290bc4c30961e4b1b4
Reviewed-on: https://gerrit.franz.com:9080/1610
Reviewed-by: John O'Rourke <jor@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Tested-by: Kevin Layer <layer@franz.com>