bug20531 Hide connection pool error when destroying
tests-added:   AGConnPoolTest
tests-run:     prepush, test-bigger
performance:   no impact

<release-notes>
bug20531 Hide connection pool error when destroying

When an AGraph server has session timed out,
and the client-side pool closes the connection,
the exception will now not be logged as a warning
because this is a normal, expected situation.

Also, uses size instead of ping to validate a connection.
</release-notes>

Change-Id: I7187b538370371f420503e24541752c6fbc6dd7b