bug20740: use java-catalog in tests and tutorials
With this change, java and clojure tests and tutorials now use
the java-catalog instead of the java-tutorial catalog.

This change requires the corresponding change to agraph-tests.cfg
in the agraph repo.  [Previously, agraph-tests.cfg used java-tutorial
catalog for java-prepush and java-bigger-test.  With this change, it
uses java-catalog in order to be in sync with the agraph.cfg file that
is generated during an agraph installation; this will also facilitate
running java tests against an installed agraph.]

java-prepush and java-bigger-test both pass.

Change-Id: I140255fd6a0238077a36ba5135f80c0d4c98a429
Reviewed-on: https://gerrit.franz.com:9080/1910
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Tested-by: Kevin Layer <layer@franz.com>