bug21092: Clojure: errors in tutorial
tests-added   : run tutorial in prepush
tests-run     : prepush
performance   : no impact

<release-notes>
bug21092: Clojure: errors in tutorial

Enhanced repo-federation to work without rcon-args as was
used in the tutorial. This fixes tutorial example16.

Enhanced value-factory to work with either Repository
or RepositoryConnection. This reduces reflection warnings
in the tutorial.
</release-notes>

Run the clojure tutorial as part of prepush to catch errors.

Tested tutorial with 'lein repl'.

Change-Id: I5e7822c91a4e93b1f4600a5700f615ab812d08ae
Reviewed-on: https://gerrit.franz.com:9080/2217
Reviewed-by: John O'Rourke <jor@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Tested-by: Kevin Layer <layer@franz.com>