bug21126: the java events ptest is broken
tests-run     : events.sh
release-notes : none, test only

Add clojure.jar to events.sh.  This is not used at runtime of Events,
but there are class references, so it is required.  This is only
needed for cases where the classes dir is used without the agraph.jar
which already references clojure.jar by relative path.

Change-Id: Id52fb68b5637aacd3a5732a2f1cbc243601b5a50