rfe11906: support materialization
<release-note>
With this change, the Java client now supports materialization, allowing
inferences to be generated and added to a repository in a transactional
fashion.  See the Javadoc for AGMaterializer#newInstance() to acquire a
configurable materializer, and AGRepositoryConnection#materialize to run
the materializer.
<release-note>

added prepush test AGMaterializerTests
make prepush passes
make javadoc runs clean.

Change-Id: I97cbb592a100f6df4e1974fa7e1fc83b1766400c
Reviewed-on: https://gerrit.franz.com:9080/2641
Reviewed-by: John O'Rourke <jor@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Tested-by: Kevin Layer <layer@franz.com>