rfe8384: limiting the duration of a query
<release-note>
rfe8384: limiting the duration of a query

With this change, the Java client now supports limiting the duration
of a query.  When Sesame applications use the AGQuery#setMaxQueryTime
method, the AG server will timeout the query appropriately during
evaluation and clean up server resources, and the client library will
throw a QueryInterruptedException back to the client application.
</release-note>

AGRepositoryConnectionTest#testOrderByQueriesAreInterruptable passes.
make prepush passes.

Change-Id: I6aaa1906cb4c2f9817d1d93b5c0a4348ba309a69
Reviewed-on: https://gerrit.franz.com:9080/2722
Reviewed-by: John O'Rourke <jor@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Tested-by: Kevin Layer <layer@franz.com>