Get rid of use of .git-branch-name file
The hook that creates .git-branch-name is considered obsolete and does
not exist on all machines.

The .git-branch-name file was used by the ant build script to set
various variables, used for file naming and documentation purposes.
Standardize on AGVERSION in the makefile, which is passed by
build scripts from the acl repo and agraph/lisp/GNUmakefile.

Change-Id: Ic5315187318f31821fbc0c92e5cf2f0fa9ee7240