bug21953. Don't always write POST query params to body.
If the body of a POST request is not
application/x-www-form-urlencoded, an HTTP server may not search for
the query parameters in the body of the request. If the
content type is application/x-www-form-urlencoded and there is no body
to the request, write parameters to the body. Else, write them to the
URL.

Tests added for:     none
make test-suite run? java-bigger-test

<release-note>
bug21953. Fix generation of some POST method requests.

AGHTTPClient would write query parameters to the body of POST requests
when the content-type was not application/x-www-form-urlencoded. This
can result in http servers not finding the query parameters, and
processing the request incorrectly. This bug corrects the problem by
writing POST request query parameters to the URL, unless the request
has no body and the content type is application/x-www-form/urlencoded.

AGRepositoryConnection.load() now works for importing server-side
statements.

</release-note>

Change-Id: Ib5aa301fdf0c6eb716722090a0e60d002780a6ca
Reviewed-on: https://gerrit.franz.com:9080/3340
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Reviewed-by: John O'Rourke <jor@franz.com>
Tested-by: Kevin Layer <layer@franz.com>