rfe12952: update java client for UseMainPortForSessions
tests run: agtest.sh 2
tests added: no

<release-note>
rfe12952 - update java client for handling of useMainPortForSessions

When the server configuration option UseMainPortForSessions is
enabled, all requests to sessions are multiplexed through the frontend
port. In this case, the java system property useMainPortForSession is
redundant and its value is not consulted. This means that one can turn
on this feature on the client if it's not enabled in the server
config, but cannot turn it off.
</release-note>

Actually, to get the tests run, the undocumented system property

  com.franz.agraph.http.overrideServerUseMainPortForSessions

can make the useMainPortForSessions property to take precedence over
the server configuration.

Change-Id: I6bcaeb0ad209960997a77801eaf3e2f0d5da9fd5
Reviewed-on: https://gerrit.franz.com:9080/3959
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Reviewed-by: John O'Rourke <jor@franz.com>
Tested-by: Kevin Layer <layer@franz.com>