rfe12546: Add getDuplicateStatements to AGRepositoryConnection
Implements AGRepositoryConnection.getDuplicateStatements in the java
client, which returns the set of duplicate triples in the repository.
Takes a `comparisonMode` argument to select how duplicates are
determined. Acceptable values are "spog" or "spo".

Tests added for:     DeleteDuplicateTests.java
make test-suite run? java-prepush

<release-note>
rfe12546: Add getDuplicateStatements to AGRepositoryConnection

Adds a new method to the AGRepositoryConnection class called
getDuplicateStatements, that returns the set of duplicate triples in
the repository. The method signature is:

public RepositoryResult<Statement> getDuplicateStatements(String comparisonMode)
       throws RepositoryException {

Duplicate triples are determined by the `comparisonMode` argument,
which can be specified as either "spo" (compare as ntriples) or "spog"
(compare as nquads).

</release-note>

javadocs are autogenerated. Else, see comment preceeding the definition
of getDuplicateStatements() in com/franz/agraph/repository/AGRepository.java

Change-Id: Ib776c7c7cd5e082ee7d46b3b4f7a4bf801565cc3
Reviewed-on: https://gerrit.franz.com:9080/4172
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Reviewed-by: John O'Rourke <jor@franz.com>
Tested-by: Kevin Layer <layer@franz.com>