rfe13510: (test only) update tests for changes on the server side
tests added    : just modified
tests run      : make java-prepush
performance    : no change
release note   : none; test only change

This change in the tests works around the issue in rfe13511 which is
caused by changes for rfe13510.

A change I haven't committed yet makes use of query plans in the HTTP
REST query API. That change leads to plans being re-used which is good.
However, the variable check occurs during _planning_ and so the value
of setCheckVariables (or, equivalently,
*sparql-check-variable-errors-p*) doesn't matter when the same query
gets re-run. This leads to the existing Java test to fail (because it
assumes that each call to the query engine will do enough work that the
value of *sparql-check-variable-errors-p* matters). So I filed a new RFE
to make it so that *sparql-check-variable-errors-p* will be applied each
time but I also think that relying on *sparql-check-variable-errors-p*
when re-using plans isn't a common occurrence and so the fix for that
should be very low priority. ...And so, I also pushed this change to
work around the changed behavior...

Change-Id: I0564611cc9e3d592c891a395e955d4f9ce1ae553
Reviewed-on: https://gerrit.franz.com:9080/4464
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Reviewed-by: John O'Rourke <jor@franz.com>
Tested-by: Kevin Layer <layer@franz.com>