rfe14296: Create a thread to keep the session alive
The session is now pinged by a periodic task created by
a ScheduledExecutorService. By default, this service is
implemented by a single thread pool stored in AGServer.
It can be overriden when creating new repository
connections (this can be used to disable the "pinger").

Added a test to ensure that sessions do not expire
and modified the old timeout test to disable the
new feature.

Change-Id: I7cdb3a4f10e65425e4a4e7ac1128c4bdf59c8ace
Reviewed-on: https://gerrit.franz.com:9080/5565
Reviewed-by: Tadeusz Sznuk <ts@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>