rfe14418: Add user Attribute API to java-client
Adds public setUserAttributes() and getUserAttributes() methods to
the AGRepositoryConnection class, which store/retrieve
userAttributes to/from the underlying AGHTttpRepoClient class.

The set method accepts either a JSONObject or a String representing
a collection of key/value pairs. The key must be a String naming a
defined attribute. The value of each pair must be either a String,
or an array of Strings, indicating that the attribute is being set
multiple times.

While userAttributes are set, an 'x-user-attributes'
header is added to any HTTP request made using the
AGRepositoryConnection, with the userAttributes
as its value.

Tests added for:       yes, AGTripleAttributesTest,
      	    	       tutorial example25()
make test-suite run?   yes
Documentation changes? yes, javadoc

Change-Id: I6e480265983f2e74bc6bbc093ce9dfcbc8c1ce2a
Reviewed-on: https://gerrit.franz.com:9080/5897
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Tested-by: Ahmon Dancy <dancy@franz.com>