rfe14408: java-client support attributes during file import
Extends support for attributes to the client- and server-side
File Import API in the java-client. Adds a number of new add()
method signatures to AGRepositoryConnection that accept an
attributes argument when importing RDF formatted data.

Cleans up a few TODOs that were encountered in the src and
test suite.

Tests added for:       AGTripleAttributesTest,
		       AGRepositoryConnectionTests
make test-suite run?   yes
Documentation changes? javadocs

Change-Id: I19bb3205556289ac4f14fac2499df360dc7bfbcf
Reviewed-on: https://gerrit.franz.com:9080/5965
Reviewed-by: Mikel Bancroft <mikel@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>