rfe14482 Add standalone AttributesExample to java-client
Add a standalone AttributesExample demo to the tutorial demo.

Move the UserAttributesContext class out of TutorialExamples and into
the java-client proper. It is added to the com.franz.agraph.repository
package.

Update TutorialExamples so it can find the UserAttributesContext
class.

Update makefile to include AttributesExample.java in the
java-client dist.

Also, add a test in AGTripleAttributesTest.java to the Prepush test suite.
It was mistakenly left in the Temp suite for isolated testing.

Tests added for:       none
make test-suite run?   yes, java-prepush
Documentation changes? none

Change-Id: I7e20217c861b272e299f1d3431aad993a5e9820d
Reviewed-on: https://gerrit.franz.com:9080/6180
Reviewed-by: Mikel Bancroft <mikel@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>