rfe14437 Add convenience function for making database connection in Java.
Adds the following methods to the AGServer class as a convenience for
creating/opening a repository and retrieving the AGRepository or
AGRepositoryConnection instance without needing to instantiate a
number of intermediary objects.

Adds a getServer() method to the AGRepositoryConnection class as a
shortcut to fetching the AGServer object associated with the
connection.

 public AGRepository createRepository(String, AGCatalog, boolean)
 public AGRepository createRepository(String, String, boolean)
 public AGRepository createRepository(String, String)
 public AGRepository createRepository(String)
 public static AGRepository createRepository(String, String, String, String, String)
 public AGRepositoryConnection createRepositoryConnection(String, String, boolean)
 public static AGRepositoryConnection createRepositoryConnection(String, String, String, String, String)

Now, instead of writing the following code:

  AGServer server = new AGServer(server_url, login, password);
  AGCatalog cat = server.getCatalog(catalog_name);
  // check for non-existent catalog
  AGRepository repo = catalog.createRepository(repo_name);
  repo.initialize();
  AGRepositoryConnection conn = repo.getConnection();

The following can be written to retrieve the AGRepository instance

  AGRepository repo =
    AGServer.createRepository(repo_name, catalog_name, server_url, user, passwd);

Or the following, to retrieve the AGRepositoryConnection instance

  AGRepositoryConnection conn =
    AGServer.createRepositoryConnection(repo_name, catalog_name, server_url, user, passwd);

Update the tutorials to use this convenience function where appropriate.
Notably, example1() in both TutorialExamples and JenaTutorialExamples are
not changed since they are intended to demonstrate how to access and use
the intermediate AGServer, AGCatalog, and AGRepository objects.

Fixes some faulty try-finally logic in AttributeExample.java.

Adds a clear indicator that the Jena Tutorial has finished.

Some additional general code cleanup in TutorialExamples.java.

Tests added for:       none
make test-suite run?   yes, all 3 tutorials run manually.
Documentation changes? none

releases notes: non-user-visible API/code change(s)

Change-Id: Ib760030fe87b5221b50698b1d1cf9ef9961343ed
Reviewed-on: https://gerrit.franz.com:9080/6228
Reviewed-by: Mikel Bancroft <mikel@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Łukasz Jancewicz <dzinx@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>