bug24119: Pinger error when session is closed
Pinger now logs errors with 'debug' priority. Since we allow the
session to be closed e.g. by deleteing the repository, and we do
not treat that as an error during close, this is the only way to
prevent spurious error messages during tests/tutorial.

Also removed two static imports.

Change-Id: I5f59851d5ecf29cece83b0db417519863b53b503
Reviewed-on: https://gerrit.franz.com:9080/6264
Reviewed-by: Tadeusz Sznuk <ts@franz.com>
Tested-by: Hudson <hudson@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>