Fix problems with the release process
The release scripts expect 'make dist CUSTOMER_DIST=y' to place
an archive in the DIST directory. Makefile in v6.2.0 used not to
do this. It now creates the expected file and makes sure that all
files in the archive are makred as owned by root (another thing
that v6.1 does when CUSTOMER_DIST is defined).

Change-Id: I8c93efe72f6c1b8afbb4430b495749784e2cbfb1
Reviewed-on: https://gerrit.franz.com:9080/6799
Reviewed-by: Tadeusz Sznuk <ts@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Tested-by: Ahmon Dancy <dancy@franz.com>